rtorrent-session-rewrite/requirements.txt
python-bencode==1.0.2
click==6.6
